requirements.txt
openai==0.28
python-dotenv
flask
boto3
streamlit